+++ javers-spring-jpa/src/test/java/org/javers/spring/example/JaversSpringJpaApplicationConfigExample.java
@@ -9,7 +9,7 @@ import org.javers.repository.sql.DialectName;
 import org.javers.repository.sql.JaversSqlRepository;
 import org.javers.repository.sql.SqlRepositoryBuilder;
 import org.javers.spring.auditable.AdvancedCommitPropertiesProvider;
-import org.javers.spring.auditable.AuditingExecutionContext;
+import org.javers.spring.auditable.AuditedMethodExecutionContext;
 import org.javers.spring.auditable.AuthorProvider;
 import org.javers.spring.auditable.CommitPropertiesProvider;
 import org.javers.spring.auditable.SpringSecurityAuthorProvider;
@@ -120,7 +120,7 @@ public class JaversSpringJpaApplicationConfigExample {
         return new AdvancedCommitPropertiesProvider() {
 
             @Override
-            public Map<String, String> provideForCommittedObject(AuditingExecutionContext ctx, Object domainObject) {
+            public Map<String, String> provideForCommittedObject(AuditedMethodExecutionContext ctx, Object domainObject) {
                 return Map.of(
                     "TargetMethodName", ctx.getTargetMethodName(),
                     "TargetClassName", ctx.getTargetClassName()
@@ -128,7 +128,7 @@ public class JaversSpringJpaApplicationConfigExample {
             }
 
             @Override
-            public Map<String, String> provideForDeletedObject(AuditingExecutionContext ctx, Object domainObject) {
+            public Map<String, String> provideForDeletedObject(AuditedMethodExecutionContext ctx, Object domainObject) {
                 return Map.of(
                     "TargetMethodName", ctx.getTargetMethodName(),
                     "TargetClassName", ctx.getTargetClassName()
@@ -136,7 +136,7 @@ public class JaversSpringJpaApplicationConfigExample {
             }
 
             @Override
-            public Map<String, String> provideForDeleteById(AuditingExecutionContext ctx, Class<?> domainObjectClass, Object domainObjectId) {
+            public Map<String, String> provideForDeleteById(AuditedMethodExecutionContext ctx, Class<?> domainObjectClass, Object domainObjectId) {
                 return Map.of(
                     "TargetMethodName", ctx.getTargetMethodName(),
                     "getTargetClassName", ctx.getTargetClassName()
