+++ javers-spring/src/main/java/org/javers/spring/auditable/AdvancedCommitPropertiesProvider.java
@@ -1,71 +1,113 @@
 package org.javers.spring.auditable;
 
+import org.javers.core.Javers;
+import org.javers.repository.api.JaversRepository;
+import org.javers.spring.annotation.JaversSpringDataAuditable;
+import org.springframework.data.repository.CrudRepository;
+
 import java.util.Collections;
 import java.util.Map;
 
 /**
- * An advanced version of {@link CommitPropertiesProvider} that supplies commit properties for Javers during auditing.
- * <p>
- * By implementing this interface, you can leverage information from the {@link AuditingExecutionContext}
- * to generate commit properties that are specific to the current operation being performed.
- * </p>
+ * A new version of {@link CommitPropertiesProvider}.
+ * <br/><br/>
+ *
+ * Use it to provide commit properties
+ * for {@link Javers#commit(String, Object, Map)}
+ * called by the auto-audit aspect &mdash; {@link JaversSpringDataAuditable}.
+ * Implementation has to be thread-safe.
+ * <br/><br/>
+ *
+ * By implementing this interface, you can leverage information from {@link AuditedMethodExecutionContext}
+ * to generate commit properties that might be aware of a current operation being performed.
+ * <br/><br/>
+ *
+ * <b>Usage</b>
+ * <br/>
+ * Create a bean in your Spring context, for example:
+ * <pre>
+ * {@code
+ *     @Bean
+ *     @ConditionalOnMissingBean
+ *     public AdvancedCommitPropertiesProvider advancedCommitPropertiesProvider() {
+ *         return new MyAdvancedCommitPropertiesProvider();
+ *     }
+ * }
+ * </pre>
  *
  * @author Xiangcheng Kuo
  * @see CommitPropertiesProvider
- * @see AuditingExecutionContext
- * @see org.javers.core.Javers
- * @since 2024-06-07
+ * @see AuditedMethodExecutionContext
+ * @since 7.5
  */
 public interface AdvancedCommitPropertiesProvider {
 
 	/**
-	 * Provides commit properties specific to the object being saved.
+	 * Provides Javers commit properties when a given object is committed (saved or updated)
+	 * to {@link JaversRepository}.
+	 * <br/><br/>
 	 *
-	 * @param ctx          the auditing context
-	 * @param domainObject the object being saved
+	 * This method is called by the {@link JaversSpringDataAuditable} aspect
+	 * to get properties for Javers commit created when
+	 * {@link CrudRepository#save(Object)} and
+	 * {@link CrudRepository#saveAll(Iterable)} methods are called.
+	 *
+	 * @param ctx          an audited method call context
+	 * @param domainObject an object being saved
 	 * @return a map of commit properties
 	 */
-	default Map<String, String> provideForCommittedObject(AuditingExecutionContext ctx, Object domainObject) {
+	default Map<String, String> provideForCommittedObject(AuditedMethodExecutionContext ctx, Object domainObject) {
 		return Collections.emptyMap();
 	}
 
 	/**
-	 * Provides commit properties specific to the object being deleted.
+	 * Provides Javers commit properties when a given object is deleted from {@link JaversRepository}.
+	 * <br/><br/>
 	 *
-	 * @param ctx          the auditing context
-	 * @param domainObject the object being deleted
+	 * This method is called by {@link JaversSpringDataAuditable} aspect
+	 * to get properties for Javers commit created when
+	 * {@link CrudRepository#delete(Object)} and
+	 * {@link CrudRepository#deleteAll(Iterable)} methods are called.
+	 *
+	 * <br/><br/>
+	 * Default implementation delegates to {@link #provideForCommittedObject(AuditedMethodExecutionContext, Object)}
+	 *
+	 * @param ctx          an audited method call context
+	 * @param domainObject an object being deleted
 	 * @return a map of commit properties
 	 */
-	default Map<String, String> provideForDeletedObject(AuditingExecutionContext ctx, Object domainObject) {
+	default Map<String, String> provideForDeletedObject(AuditedMethodExecutionContext ctx, Object domainObject) {
 		return Collections.emptyMap();
 	}
 
 	/**
-	 * Provides commit properties specific to the object being deleted by its ID.
+	 * Provides Javers commit properties when a given object is deleted from {@link JaversRepository}
+	 * by its Id.
+	 * <br/><br/>
 	 *
-	 * @param ctx               the auditing context
-	 * @param domainObjectClass the class of the object being deleted
-	 * @param domainObjectId    the ID of the object being deleted
+	 * This method is called by {@link JaversSpringDataAuditable} aspect
+	 * to get properties for Javers commit created when
+	 * {@link CrudRepository#deleteById(Object)} is called.
+	 *
+	 * <br/><br/>
+	 * Default implementation returns empty Map
+	 *
+	 * @param ctx               an audited method call context
+	 * @param domainObjectClass a class of the object being deleted
+	 * @param domainObjectId    an ID of the object being deleted
 	 * @return a map of commit properties
 	 */
-	default Map<String, String> provideForDeleteById(AuditingExecutionContext ctx, Class<?> domainObjectClass, Object domainObjectId) {
+	default Map<String, String> provideForDeleteById(AuditedMethodExecutionContext ctx, Class<?> domainObjectClass, Object domainObjectId) {
 		return Collections.emptyMap();
 	}
 
 	/**
-	 * Returns an instance of an empty implementation of {@code AdvancedCommitPropertiesProvider}.
-	 *
-	 * @return an instance of {@code EmptyAdvancedCommitPropertiesProvider}
+	 * Default implementation returning empty maps.
 	 */
 	static AdvancedCommitPropertiesProvider empty() {
 		return new EmptyAdvancedCommitPropertiesProvider();
 	}
 
-	/**
-	 * An empty implementation of {@code AdvancedCommitPropertiesProvider} that returns empty maps for all methods.
-	 */
 	class EmptyAdvancedCommitPropertiesProvider implements AdvancedCommitPropertiesProvider {
-
 	}
-
 }
