+++ javers-spring/src/main/java/org/javers/spring/auditable/aspect/springdata/AbstractSpringAuditableRepositoryAspect.java
@@ -6,7 +6,7 @@ import org.javers.repository.jql.QueryBuilder;
 import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.javers.spring.auditable.AdvancedCommitPropertiesProvider;
 import org.javers.spring.auditable.AspectUtil;
-import org.javers.spring.auditable.AuditingExecutionContext;
+import org.javers.spring.auditable.AuditedMethodExecutionContext;
 import org.javers.spring.auditable.AuthorProvider;
 import org.javers.spring.auditable.CommitPropertiesProvider;
 import org.javers.spring.auditable.aspect.JaversCommitAdvice;
@@ -26,13 +26,13 @@ public class AbstractSpringAuditableRepositoryAspect {
     }
 
     protected void onSave(JoinPoint pjp, Object returnedObject) {
-        AuditingExecutionContext ctx = AuditingExecutionContext.from(pjp);
+        AuditedMethodExecutionContext ctx = AuditedMethodExecutionContext.from(pjp);
         getRepositoryInterface(pjp).ifPresent(i ->
                 AspectUtil.collectReturnedObjects(returnedObject).forEach(it -> javersCommitAdvice.commitObject(ctx, it)));
     }
 
     protected void onDelete(JoinPoint pjp) {
-        AuditingExecutionContext ctx = AuditingExecutionContext.from(pjp);
+        AuditedMethodExecutionContext ctx = AuditedMethodExecutionContext.from(pjp);
         getRepositoryInterface(pjp).ifPresent( i -> {
             RepositoryMetadata metadata = DefaultRepositoryMetadata.getMetadata(i);
             for (Object deletedObject : AspectUtil.collectArguments(pjp)) {
@@ -51,7 +51,7 @@ public class AbstractSpringAuditableRepositoryAspect {
     }
 
 
-    void handleDelete(AuditingExecutionContext ctx, RepositoryMetadata repositoryMetadata, Object domainObjectOrId) {
+    void handleDelete(AuditedMethodExecutionContext ctx, RepositoryMetadata repositoryMetadata, Object domainObjectOrId) {
             if (isIdClass(repositoryMetadata, domainObjectOrId)) {
                 Class<?> domainType = repositoryMetadata.getDomainType();
                 if (javers.findSnapshots(QueryBuilder.byInstanceId(domainObjectOrId, domainType).limit(1).build()).size() == 0) {
